Island: Move register_event_handlers to its own module
